feat: add dedicated admin interface
